Make all tests pass.
Add EasyMock method expectations for the findAll(),
findByLabel(), and findByForeignId() methods.
Add a couple of nodes with different foreign sources and Ids.

Remove the calls to save() and saveOrUpdate() in the unit test,
and instead test with nodes that are created by populateNodes.setUpMock().

Change the dirtiesContext property from false to true.

Add a parm to the test events with the foreignSource that should be used
for the findByForeignId() method call.